% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{inv_logit}
\alias{inv_logit}
\title{inv_logit: Returns the inverse logit, or "Sigmoid" of the input}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Input}
}
\value{
inv_logit(x), or equivalently, Sigmoid(x)
}
\description{
inv_logit: Returns the inverse logit, or "Sigmoid" of the input
}
\references{
\url{https://github.com/DavidKLim/dlglm}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
