% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{prepareData}
\alias{prepareData}
\title{prepareData: Main simulation wrapper function}
\usage{
prepareData(
  dataset = "SIM",
  data.file.name = NULL,
  mask.file.name = NULL,
  sim.params = list(N = 1e+05, D = 2, P = 8, data_types_x = NA, family = "Gaussian",
    sim_index = 1, ratios = c(train = 0.8, valid = 0.1, test = 0.1), beta = 0.25, C =
    NULL, Cy = NULL, NL_x = F, NL_y = F),
  miss.params = list(scheme = "UV", mechanism = "MNAR", pi = 0.5, phi0 = 5,
    miss_pct_features = 50, miss_cols = NULL, ref_cols = NULL, NL_r = F),
  case = c("x", "y", "xy")
)
}
\arguments{
\item{dataset}{Dataset name (default: "SIM"). Inference and prediction L1 computed just for SIM}

\item{data.file.name}{Default is NULL. Can load custom data file name with sim.data, X, and Y}

\item{mask.file.name}{Default is NULL. Can load custom missingness file name with sim.mask, mask_x, and mask_y}

\item{sim.params}{List of simulation parameters. Contains N, D, P, data_types_x, family, sim_index, ratios (split of train, valid, test)}

\item{miss.params}{List of missingness parameters. Contains scheme, mechanism, pi, phi0, miss_pct_features, miss_cols, ref_cols, NL_r}

\item{case}{Simulation case of missingness (missingness in x, y, or both xy)}
}
\value{
Saves and returns list of simulation parameters and data: X, Y, mask_x, mask_y, g, sim.params, miss.params, sim.data, sim.mas
}
\description{
prepareData: Main simulation wrapper function
}
\examples{
EXAMPLE HERE

}
\references{
\url{https://github.com/DavidKLim/dlglm}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
