% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processResults.R
\name{processResults}
\alias{processResults}
\title{processResults: Processes final dlglm, idlglm, and mice results}
\usage{
processResults(
  dataset = "SIM",
  prefix = "",
  data.file.name = NULL,
  mask.file.name = NULL,
  sim.params = list(N = 1e+05, D = 2, P = 8, family = "Gaussian", sim_index = 1, ratios
    = c(train = 0.6, valid = 0.2, test = 0.2), mu = 0, sd = 1, beta = 5, C = 3, Cy =
    NULL, NL_x = F, NL_y = F),
  miss.params = list(scheme = "UV", mechanism = "MNAR", pi = 0.5, phi0 = 5,
    miss_pct_features = 50, miss_cols = NULL, ref_cols = NULL, NL_r = F),
  case = c("x", "y", "xy"),
  data_types_x,
  data_type_y = "real",
  family = "Gaussian",
  methods = c("idlglm", "dlglm", "mice", "zero", "mean"),
  init_r = "default",
  normalize = F
)
}
\arguments{
\item{dataset}{Dataset name (default: "SIM"). Inference and prediction L1 computed just for SIM}

\item{prefix}{Prefix for dataset (default: "").}

\item{data.file.name}{Default is NULL. Can load custom data file name with sim.data, X, and Y}

\item{mask.file.name}{Default is NULL. Can load custom missingness file name with sim.mask, mask_x, and mask_y}

\item{sim.params}{List of simulation parameters. Contains N, D, P, data_types_x, family, sim_index, ratios (split of train, valid, test)}

\item{miss.params}{List of missingness parameters. Contains scheme, mechanism, pi, phi0, miss_pct_features, miss_cols, ref_cols, NL_r}

\item{case}{Simulation case of missingness (missingness in x, y, or both xy)}

\item{data_types_x}{Vector of data types ('real', 'count', 'pos', 'cat')}
}
\value{
res object: with imputation ("imp"), inference ("inf"), and prediction ("pred") metrics
}
\description{
processResults: Processes final dlglm, idlglm, and mice results
}
\examples{
EXAMPLE HERE

}
\references{
\url{https://github.com/DavidKLim/dlglm}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
