% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateData}
\alias{simulateData}
\title{simulateData: Simulates covariates matrix X and response Y}
\usage{
simulateData(
  dataset = "SIM",
  N,
  D,
  P,
  data_types_x,
  family,
  seed,
  NL_x,
  NL_y,
  beta = 0.25,
  C = NULL,
  Cy = 1
)
}
\arguments{
\item{dataset}{Dataset name (default: "SIM"). Inference and prediction L1 computed just for SIM}

\item{N}{Number of observations in data}

\item{D}{Dimensionality of simulated latent variable}

\item{P}{Number of features in data}

\item{data_types_x}{P-length vector of types of X ("real", "pos", "cat", "count")}

\item{family}{Family of response variable: "Gaussian", "Multinomial", or "Poisson" (not tested)}

\item{seed}{Seed to generate reproducible data}

\item{NL_x}{Boolean: nonlinear generation of X from Z}

\item{NL_y}{Boolean: nonlinear generation of Y from X}

\item{beta}{Base value of simulated coefficients for generation of Y from X (default: 0.25)}

\item{C}{Number of categories of "cat" variables in X. All categorical variables in X are simulated with the same number of categories}

\item{Cy}{Number of categories of categorical variable Y (family = "Multinomial")}
}
\value{
list with data and params
}
\description{
simulateData: Simulates covariates matrix X and response Y
}
\examples{
EXAMPLE HERE

}
\references{
\url{https://github.com/DavidKLim/dlglm}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
