% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{simulateMask}
\alias{simulateMask}
\title{simulateMask: Simulates missingness mask on X, based on specified mechanism and other parameters, like phi}
\usage{
simulateMask(
  data,
  scheme,
  mechanism,
  NL_r,
  pi,
  phis,
  miss_cols,
  ref_cols,
  seed
)
}
\arguments{
\item{data}{Data matrix X used as covariates for simulating the missingness}

\item{scheme}{UV (default) or MV: Univariate or multivariate missingness model}

\item{mechanism}{Mechanism of missingness ("MCAR", "MAR", or "MNAR")}

\item{NL_r}{Boolean: nonlinear generation of R from covariate(s)}

\item{pi}{Proportion of missingness in each partially-observed column, ID'ed by miss_cols}

\item{phis}{Coefficient of missingness model for each feature in X}

\item{miss_cols}{Indices of columns containing missingness (partially-observed features in X)}

\item{ref_cols}{Indices of columns not containing missingness (fully-observed features in X)}

\item{seed}{Seed to generate reproducible data}
}
\value{
list with Missing: simulated mask of X, probs: probability of each entry in X being observed, params: phis, miss_cols, ref_cols, scheme
}
\description{
simulateMask: Simulates missingness mask on X, based on specified mechanism and other parameters, like phi
}
\examples{
EXAMPLE HERE

}
\references{
\url{https://github.com/DavidKLim/dlglm}
}
\author{
David K. Lim, \email{deelim@live.unc.edu}
}
